src/corellium/requirements-pip-appium.txt
Appium-Python-Client
